Resolve Issue #36 and Update Timesheet Consistency on Time Entry Operations

+++ backend/model/repository/time_entry_repository.py
@@ -52,7 +52,6 @@ class TimeEntryRepository:
         time_entries = self.db.timeEntries.find({"date": date, "username": username})
         return list(time_entries)
 
-
     #TODO: This method is only used for testing and shouldn't be required
     def get_time_entries(self):
         """
@@ -88,29 +87,36 @@ class TimeEntryRepository:
         if result.matched_count == 0:
             return RequestResult(False, "Entry not found", 404)
         if result.modified_count == 0:
-            return RequestResult(False, "Entry update failed", 500)
+            return RequestResult(False, "Entry update failed (Not Modified)", 500)
         if result.acknowledged:
             return RequestResult(True, "Entry updated successfully", 200)
         return RequestResult(False, "Entry update failed", 500)
 
     def delete_time_entry(self, entry_id: str):
         """
-        Deletes a TimeEntry object from the database
+        Deletes a TimeEntry object from the database and returns the corresponding timesheet ID
         :param entry_id: The ID of the TimeEntry object to delete
         :return: A RequestResult object indicating the success of the operation
         """
         if entry_id is None:
             return RequestResult(False, "Entry ID is None", 400)
+
+        time_entry = self.get_time_entry_by_id(entry_id)
+        if not time_entry:
+            return RequestResult(False, "Time Entry not found", 404)
+
+        timesheet_id = time_entry.get("timesheetId")
+
         result = self.db.timeEntries.delete_one({"_id": ObjectId(entry_id)})
         if result.deleted_count == 0:
             return RequestResult(False, "Entry not found", 404)
         if result.acknowledged:
-            return RequestResult(True, "Entry deleted successfully", 200)
+            return RequestResult(True, "Entry deleted successfully", 200, data={"timesheetId": timesheet_id})
         return RequestResult(False, "Entry deletion failed", 500)
 
     def create_time_entry(self, time_entry: TimeEntry):
         """
-        Creates a new TimeEntry object in the database
+        Creates a new TimeEntry object in the database and returns the id of the entry
         :param time_entry: The TimeEntry object to create
         :return: A RequestResult object indicating the success of the operation
         """
@@ -119,9 +125,11 @@ class TimeEntryRepository:
 
         if self.get_time_entry_by_id(time_entry.time_entry_id):
             return RequestResult(False, "Time entry already exists", 409)
+
         time_entry_dict = time_entry.to_dict()
-        time_entry_dict.pop('timeEntryId')
         result = self.db.timeEntries.insert_one(time_entry_dict)
+
         if result.acknowledged:
-            return RequestResult(True, f'Time entry created successfully with ID: {str(result.inserted_id)}', 201)
+            return RequestResult(True, f'Time entry created successfully with ID: {str(result.inserted_id)}', 201,
+                                 data={"_id": ObjectId(result.inserted_id)})
         return RequestResult(False, "Time entry creation failed", 500)
